Add additional explanation in readme
In order to make the library easier to integrate into projects, this
commit updates the readme with more explanation for integration. It also
adds an additional optional method to the Configurable protocol. This is
because the default implementation for
`configurationValueForService:key:` provided by the extension on
Configurable is not visible to Objective-C code. So in order to allow
better Objective-C integration, the optional method provides a means for
Objective-C code to pull service specific values from the object that
conforms to the Configurable protocol.