Add support for Swift Package Manager
In order to support Swift Package Manager, this commit adds a couple new
files. The first is the Package.swift which declares our swift package.
The second is a directory containing a symlink to our actual codebase.
This is necessary in order to support both the traditional directory
structure that Xcode works with and the expected directory structure
that Swift Package Manager wants.